% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sb_file_list}
\alias{sb_file_list}
\title{Get a list of files associate with a project}
\usage{
sb_file_list(owner, project)
}
\arguments{
\item{owner}{The project owner}

\item{project}{The project id}
}
\description{
Get a list of files associate with a project
}

