% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sb_GET}
\alias{sb_GET}
\title{Perform an API GET request from SevenBridged CGC api}
\usage{
sb_GET(path, ..., token = sb_token())
}
\arguments{
\item{path}{path for the API call}

\item{...}{passed to the httr GET}

\item{token}{the auth token for the API. See \link{sb_token} for details.}
}
\description{
Perform an API GET request from SevenBridged CGC api
}

