% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{sb_project_details}
\alias{sb_project_details}
\title{Get project details}
\usage{
sb_project_details(owner, project)
}
\arguments{
\item{owner}{the SevenBridges owner name}

\item{project}{the name of the SevenBridges project}
}
\description{
Get project details
}

